pg-cdc-to-bq-streaming/pg-cdc-listener/requirements.txt
Flask
gunicorn
Werkzeug
pg8000 
cloud-sql-python-connector
psycopg2-binary
google-cloud-pubsub